ADD,EDIT and DELETE products ONLY for authorized users

+++ src/main/java/com/nazarov/shop/web/servlets/DeleteProductServlet.java
@@ -5,22 +5,29 @@ import com.nazarov.shop.service.ProductService;
 import javax.servlet.*;
 import javax.servlet.http.*;
 import java.io.IOException;
+import java.util.List;
 
 public class DeleteProductServlet extends HttpServlet {
-
+    private final List<String> userTokens;
     private ProductService productService;
+    public SecurityAuthChecker securityAuthChecker = new SecurityAuthChecker();
 
-    public DeleteProductServlet(ProductService productService) {
+    public DeleteProductServlet(ProductService productService, List<String> userTokens) {
         this.productService = productService;
+        this.userTokens = userTokens;
     }
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String pathInformation = request.getPathInfo();
-        int index = pathInformation.lastIndexOf("/");
-        int id = Integer.valueOf(pathInformation.substring(index + 1, pathInformation.length()));
-        productService.delete(id);
-        System.out.println("DELETE product with id: " + id);
-        response.sendRedirect("/products");
+        if (securityAuthChecker.checkUserAuthorization(request, userTokens)) {
+            String pathInformation = request.getPathInfo();
+            int index = pathInformation.lastIndexOf("/");
+            int id = Integer.valueOf(pathInformation.substring(index + 1, pathInformation.length()));
+            productService.delete(id);
+            System.out.println("DELETE product with id: " + id);
+            response.sendRedirect("/products");
+        } else {
+            response.sendRedirect("/login");
+        }
     }
 }
